Move backfill scripts out of lib/database into website/src/scripts.
